xdata/xodtemplate: Add some missing error messages
It appears as though processing group names could earlier lead to
errors when '*groups' directives were null (and void) in member
directives, and we didn't report where the error was, which is
very unfriendly, since users may have fairly huge configurations
to dig through to find the error.

This patch adds the missing error messages.

Again, many thanks to Daniel Wittenberg for assistance in debugging
this. He's now a certified gdb virtuoso ;)

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2156 5f96b256-904b-4d8d-8c98-d829582c6739