core: Fix a small compiler warning
When we have gnu getopt and are compiling with -ansi, we would get a
warning about declaration after statement. We fix that here, and
reduce the loc-count a little with a simple getopt() macro that
wraps getoptlong if it's available. This reduces the possibilities
of maintenance errors somewhat as there's only one optstring to take
care of.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2157 5f96b256-904b-4d8d-8c98-d829582c6739