core: Make precaching work without -v or -s
The two execution paths for scheduling tests and config verification
are almost identical in setup, so by merging them and joining the
precache option path to them, we can get precaching to work without
having to swim through additional output. We will get error messages
though, and now we don't have to duplicate them for either of the
codepaths.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2158 5f96b256-904b-4d8d-8c98-d829582c6739