xdata/xodtemplate: Unify slave object free()ing
Most slave objects (escalations and dependencies) had their variables
free()'d at the end, but it appears we free()'d the hostescalation
host_name and hostgroup_name twice, and we left a lot of dead memory
lying around when we didn't need to. With this patch, we unify the
logic and are once again free from "invalid free() / delete[]" from
valgrind.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2159 5f96b256-904b-4d8d-8c98-d829582c6739