Clean up global variables somewhat
The extemely lengthy lists of 'extern this and that' at the top of
every file was error prone and hard to maintain. Having the variable
definitions stored in base/nagios.c also made it nigh-impossible to
link the sources to another main() routine without creating several
stub variables that were never used.

The largest common denominator for core-only variables seems to be
base/utils.c (since that's where reset_variables() lives), so we
move the definitions there, and put core-only declarations in
include/nagios.h. Variables shared between core and cgi's are
defined in common/shared.c and declared in include/common.h.

The same work could be done for cgi's, but the savings would be
considerably smaller there, so I refrained.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2168 5f96b256-904b-4d8d-8c98-d829582c6739