Fix adding timeranges to timeperiods
Previously we tried to be clever and skip the first timeperiod entry
when we knew we couldn't break out of the loop early, but that failed
pretty hard when multiple timeranges were given in sorted order, since
we never even entered the loop in such cases.

This patch corrects it and makes sure we run the loop at least once
for every non-first timerange we add.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2169 5f96b256-904b-4d8d-8c98-d829582c6739