core: Suppress printing objectfiles unless -v is given at least twice
It gets quite annoying when trying to do timing-debug of the config
and a list of 940-ish config files keep zooming past the screen for
every run. Since it's fairly useless information for most people, we
suppress it unless -v is given multiple times.

The patch looks bigger than it is, since the 'x == TRUE' idiom only
works with booleans and not counters.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2171 5f96b256-904b-4d8d-8c98-d829582c6739