xdata: Avoid skiplisting escalations and dependencies
Doing so only takes extra time for the microscopic gain of getting
them printed in 100% alphabetical order in the objects.cache file.
Since we already attach them to their respective master objects, we
can utilize that sorting to print them in "almost-alphabetical" order
anyway. The alpha-sorting is currently only done for cgi/config.c,
but could easily be implemented for the object cache files too.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2176 5f96b256-904b-4d8d-8c98-d829582c6739