Add lib/nsock: The world's smallest socket library
This only has one function so far; nsock_unix().

The nsock_unix() function creates or connects to a unix domain
socket, depending on the flags given to it. It handleds both
udp and tcp (but not raw or seqpacket sockets), so it's more or
less fully covered for both ends of unix domain socket ipc
stuff, which is what most modules use to talk to their daemons
or with other addons.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2182 5f96b256-904b-4d8d-8c98-d829582c6739