common/objects: Make 'address' optional for hosts
Many people put fully qualified dns names as host_names and
don't really need or want to put in an extra 'address' field.
We humor them with this patch, assigning the host_name to the
address field in case address isn't set.

While at it, we do the same for 'alias', which was previously
copied (rather than assigned), and naturally make sure that
all variables are free()'d safely or ignored correctly at cleanup
time.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2183 5f96b256-904b-4d8d-8c98-d829582c6739