core: Move the timing_point() calls somewhat
The previous output was somewhat confusing, as the time used
didn't show up on the line where the time was actually spent. By
moving the timing points to after the things they time, we get a
picture that's easier to understand.

While we're at it, we also print the number of objects we added
at each point.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2186 5f96b256-904b-4d8d-8c98-d829582c6739