xdata/xodtemplate: Start counting escalations too
Since escalations can avoid skiplist insertions just as dependencies,
we have to count them manually as we duplicate them. We also add the
missing code to cound objects properly for the cgi's, which have been
a second class citizen up til now (actually, they still are, since I'm
not bothering with testing them until I'm done with the core).

Since we're now counting an awful lot of object types, we might as
well do so using an object_count struct and move all other counters
over to that one.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2189 5f96b256-904b-4d8d-8c98-d829582c6739