xdata/xodtemplate: Update timing_point() messages
It's nifty to see how many objects were created with respect to how
many services there are, since that provides a meaningful context for
later errors.

While at it, we also add the missing message for when servicegroup
recombobulation is done and fix service counting so it's never tainted
by future cgi code.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2193 5f96b256-904b-4d8d-8c98-d829582c6739