core: Speed up group registration considerable
Previously we'd engage in insertion-sorting the members into their
respective lists. That's fine and dandy when there's about five of
them, but with the massive member-lists found in large configs today
it's sheer murder to use an O(n^2) algorithm to do sorting. Until we
can be bothered to do an O(n lg n) algorithm for an unbound number of
members (heap-sort should work well), we simply stash the objects to
the front of the member list if large installation tweaks are enabled.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2194 5f96b256-904b-4d8d-8c98-d829582c6739