common/objects: Add prepend_unique_object_to_objectlist()
This is quite nifty when adding host and servicedependencies, which
can currently be duplicated (with identical values) any number of
times by having hosts join multiple hostgroups and assigning
dependencies to each of them. It might be too slow to bother with,
but services normally don't depend on more than one other service,
so it should be fairly quick.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2199 5f96b256-904b-4d8d-8c98-d829582c6739