xdata/xodtemplate: Ignore 'n' option dependencies
We don't create dependencies that are missing one (or more) of the
listed options, but we do create dependencies that have the 'n'
option as its sole failure criteria, which is essentially the same
as no criteria, except it's more verbose and more likely that a user
has no idea it will sit in RAM and consume memory while doing
absolutely nothing. With this patch, such dependencies will no longer
be created. This lets us save a smallish amount of time during config
loading.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2200 5f96b256-904b-4d8d-8c98-d829582c6739