xdata/xodtemplate: Register dependencies ad-hoc
By registering dependencies as we encounter them, we can ensure
that there are no duplicates and that we never waste memory storing
any duplicates. This comes at the expense of us having to allocate
the objects on the fly (which is slower than chunk-allocating, but a
lot faster than swapping), and not being able to assign id fields
until after all config is read (since we need to compare the deps
with each other).

There's a lot of fallout from this conceptually simple patch, so I
hope I got everything right.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2202 5f96b256-904b-4d8d-8c98-d829582c6739