Make the HOST_{UP,DOWN,UNREACHABLE} macros globally trustworthy
include/statusdata.h has earlier re-defined them with other values,
leading to no small amount of headache when doing changes that both
cgi's and core should benefit from. This patch removes the "special"
HOST_* macros and renames them SD_HOST_* instead (thank you, sed).

This also makes it a lot easier to reconcile the builds (which is
kinda low on my todo-list, but still on there), so we can get rid
of a ton of ifdef's in the code and unify a lot of the codepaths.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2218 5f96b256-904b-4d8d-8c98-d829582c6739