Remove the integers-acting-as-flags thinking
This patch has large ramifications for nebmods.

All the "fail_on_ok", "notify_on_down" fields that map naturally to
a bitgate for monitored objects' current_state variable has now been
turned into such bitgates. The bitgate is managed with the standard
paradigm "1 << enum_value", where our enum_value is "current_state"
from either a host or a service.

We provide accessor macros for handling flags properly so as to
avoid bitmath errors (or rather, to get a single place where we can
fix any such errors).

This is one step further on the road that leads to a merge of host
and service objects (and their slave objects).

The changes are added to the list in api-incompat3to4.dox.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2219 5f96b256-904b-4d8d-8c98-d829582c6739