Add service parents
They will almost certainly need some tweaking, but the basic and
very simple idea is that we will block notifications if the parent
service(s) is/are anything but OK. As with hosts; One good parent
is enough to cause notification blockage from happening.

Check scheduling is not affected in any way, and there is a whole
host of other things that probably have to be done for this to
work as people expect it, but this is good enough to be going on
with for now, as this is the change that breaks the object API
and thus has to be done before Nagios 4.

It's worth noting that we don't verify cyclic paths for services,
since we never walk that chain beyond our immediate parents.

It's also worth noting that xdata/xodtemplate.c is now an even
ungodlier mess of #ifdef NSCORE and #ifndef NSCGI. That can't
be helped, since I flat out refuse to parse service parents
manually.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2220 5f96b256-904b-4d8d-8c98-d829582c6739