xod: Fix dependency structs to match the rest of Nagios
Pretty much everything in the xod code matches the actual config
file in terms of struct member names. This patch brings the
dependency code up to scratch with that paradigm.

In itself, this patch is useless code churn, but it allows us to
make changes that simplifies adding variables to the object config.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2221 5f96b256-904b-4d8d-8c98-d829582c6739