Add check source determination capabilities
Since the introduction of mod_gearman, dnx, Merlin and worker
processes it's no longer so clear where checks come from, and users
of such systems are left pretty much in the dark about where a
certain check happened, which can be crucial to know when attempting
to find problems in a network.

With this patch, we add a "struct check_engine" which is added to
the checkresult struct, and also add a "void *source" member to the
same struct. The system is quite easy, but a bit more convoluted
than it has to be, since redundancy modules will want to know where
checks came quickly in order to know if they should send the checks
onwards or not.

With this patch that becomes both simple and efficient, and we can
provide const strings for each of our workers, and we can utilize
the same functionality to identify external special-purpose workers
when we get those, which is likely to be an amazing help during the
debugging of those.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2230 5f96b256-904b-4d8d-8c98-d829582c6739