process_passive_service_check(): Take description of looked-up service
It makes little sense to use a pointer into the iocache from the
external command buffer, as it gets overwritten from time to time.
With this patch we make things more future-proof in case things
want to stash a pointer to the service_description from the check
result struct later.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2231 5f96b256-904b-4d8d-8c98-d829582c6739