core: Refactor event handling loop somewhat
Previously we used to remove the event structure before we checked
if we were supposed to run the event, which caused the event_runtime
fetched from squeue_get_runtime() to point to invalid memory and
sometimes be filled with random crap, causing latency calculations
to be waaay off (like 50 billion years or so).

With this patch we avoid most of the if()'s and move on quicker if
we're not supposed to run the check, removing it explicitly where
needed.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2232 5f96b256-904b-4d8d-8c98-d829582c6739