Eventbroker API: Add support for built-in core modules
It's usually a good idea to eat ones own dogfood. It's also a
pretty good idea to be able to incorporate nifty modules directly
in to the core, should we decide to go that way.

With this patch, that becomes possible, since we can now add core
modules that can register callbacks and don't cause segfaults when
they're about to be unloaded, and soon that will become highly
useful.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2238 5f96b256-904b-4d8d-8c98-d829582c6739