Deprecate {HOST,SERVICE}_CHECK_{ACTIVE_PASSIVE}
Since the way things are checked doesn't differ between hosts and
services, it's just confusing to have two different macros that
mean the exact same thing.

Instead we introduce CHECK_TYPE_{ACTIVE,PASSIVE,PARENT,FILE,OTHER},
with ACTIVE and PASSIVE replacing the four previous similar ones.
CHECK_TYPE_PARENT means we've run an (active) check for the benefit
of an object that depends on the object we just checked. FILE means
we read the check result from a spooled file (which is deprecated
and should just go away as quickly as possible), and OTHER being a
catch-all to let modules set something that marks checks from them
as special.

We retain the old macros yet a while longer, but clearly mark them
as deprecated, and we get rid of them from the sources so others
don't use them by mistake.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2239 5f96b256-904b-4d8d-8c98-d829582c6739