lib/lnag-utils: Add ranged_urand()
Proper randomization is highly useful in algorithms where no good
solution exists, but it's boring to type and off-by-ones tend to
sneak in when coding at three in the morning. With this helper we
avoid such errors, which is a Good Thing(tm) since such bugs are
by their very definition impossible to reliably reproduce.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2240 5f96b256-904b-4d8d-8c98-d829582c6739