core: Run first check of objects within their max check delay
If the initially scheduled checks happen to end up too far into the
future we simply schedule them on a random slot within their
check window. The check_window() macro takes care of calculating the
max time we have, regardless of whether the host (or service) is in
soft or hard state.

This fixes issue 83.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2241 5f96b256-904b-4d8d-8c98-d829582c6739