base/checks: Stop being tricksy with orphan detection checks
The orphan detection logic might be doing it wrong with regards to
hosts (it looks correct, but that's no guarantee), but intentionally
fooling ourselves with a bug in one place to work around a bug in
another makes absolutely no sense what so ever.

This patch removes one such cover-up bug. If we get the bug with
orphaned host checks back (unlikely, as they're run by the worker
processes now), we can fix that properly instead of working around
it in such a hackish manner.

This fixes issue 270.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2243 5f96b256-904b-4d8d-8c98-d829582c6739