xod: Fix contact inheritance
According to the docs, contacts should only be inherited from host to
service if the service has no other contacts what so ever (and the
same goes for escalations), but the way the code worked was that it
handled contact_groups and contacts directives separately, meaning
services with only 'contacts' specified were still eligible for
inheriting 'contact_groups' from the host, leading to much confusion.

This patch fixes that, and thereby issue 314.

While we're at it, we re-organize the inheritance macros a bit and
make sure to use them for cross-object inheritance too.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2244 5f96b256-904b-4d8d-8c98-d829582c6739