Bring snprintf.o into libnagios
I'm not entirely certain this is a good idea, but without it it's
possible the library won't have access to vasprintf() and other
nifty string builders, so we do it anyway.

"It works on my system" isn't very reassuring here, since
snprintf.o is more or less empty on every modern system I can
think of. Thankfully though, modern systems (well, modern enough
to obey one of the simplest parts of C99) is our primary target
audience. We may have to fix it later though.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2248 5f96b256-904b-4d8d-8c98-d829582c6739