Properly handle check output when stdout is empty and stderr is not
The key/value vector we get from the worker always has values
for everything, so we must check for the nul string rather than
NULL when checking if there is output on stdout. This sets the
output from the check to this message:

 (No output on stdout) stderr: execvpe() failed. errno is %d: %s

where %s is ofcourse strerror(errno), usually letting people know
that the file they're trying to execute is missing, or that they
lack permissions to run the check.

Reported-by: Joerg Linge <pitchfork@ederdrom.de>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2250 5f96b256-904b-4d8d-8c98-d829582c6739