lib/iobroker: Make it an error to re-register a filedescriptor
Without this patch we would have leaked memory and silently
overwritten a previously registered handler when someone
registered a new handler for the same filedescriptor. Since
multiple input handlers for a single source makes no sense (and
weren't supported earlier either), we simply error out when we
notice that happening.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2257 5f96b256-904b-4d8d-8c98-d829582c6739