core: Re-use pre-existing timed_event's for hosts and services
This fixes an extremely intermittent bug which would previously
occur when we happened to get an inbound checkresult from a worker
for an on-demand check of a host while inspecting the timed_event
belonging to that host.

When such a host received a new scheduled item, they used to free()
their current timed_event, only to allocate a new one immediately.
That didn't make much sense, cost us performance, caused the event
execution loop to crash and lost me a full day of hacking on the
general-purpose all-magical superduper ultra-fantastic query handler,
now with extra bling.

As an extra fun thing, it appears this can (well, could) only happen
before all hosts and/or services had been checked at least once, and
only for hosts, and almost certainly only with the pathological check
settings I'm using for testing (2 second interval_length), and most
likely only when the developer is really, really tired of looking at
valgrind (which caused the bug NOT to happen for 900 invocations) and
quite possibly only on rainy wednesdays between 12.30 and midnight...

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2261 5f96b256-904b-4d8d-8c98-d829582c6739