base/events: Clean up event handling a smidgeon
Most of these changes were deemed necessary during The Hunt For The
Horrible Bug four commits back. It's a tad cleaner and a bit less
code now, so I assume that's good. We also stop removing scheduled
service checks twice. Many thanks and kudos to Armin Wolfermann for
spotting that and sending the first patch, which caused me to take
a deeper look.

Inspired-by: Armin Wolfermann <aw@osn.de>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2266 5f96b256-904b-4d8d-8c98-d829582c6739