nerd: Add opathchecks channel
The output from a subscription here can be piped to gource to draw
a real-time fly-around whizzbang graph of how checkresults come in
from various sources. It's pretty damn cool, even if I do say so
myself.

We still don't support state-based colors, but adding it is a very
quick hack indeed, so it won't be long.

Signed-off-by: Andreas Ericsson <ae@op5.se>

Conflicts:
	base/nerd.c

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2275 5f96b256-904b-4d8d-8c98-d829582c6739