core: Fix hostdependency checking
Previously we tried to use a NULL pointer to check if we should
add hosts to the check-list when using predictive dependency
checking. The compiler failed to warn us about that, since we
explicitly initialized the temp_hostdependency pointer to NULL
and therefore didn't use it as an unassigned value.

We fix it long-term by moving the declaration to the usage
point and making sure we initialize it.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2281 5f96b256-904b-4d8d-8c98-d829582c6739