Undo bulk-allocation of objects
If we do bulk-allocation, adding objects during runtime becomes a
herculean task. We want things to be simple, so we keep the *_list
lists (err...) as arrays of pointers and the addresses of individual
objects immutable.

On the downside, this is a huge change. On the upside, this makes
post-processing object config simpler.

Suggested-by: Daniel Wittenberg <dwittenberg2008@gmail.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2284 5f96b256-904b-4d8d-8c98-d829582c6739