Undo the null dereference list-to-pointer-array craziness
Coding while drunk is apparently not always such a stellar idea.

Instead, we add *_ary pointer-to-pointer arrays for all objects
so we have some means of getting hold of the objects that lack
'next' pointers (which will be all of them in Nagios 5).

This makes porting modules from Nagios 4 to Nagios 5 a lot more
pleasant, avoids the null pointer dereference and removes the
hugely ugly and unwieldy ternary statements that protected us
against coredumps caused by some object type not being present.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2289 5f96b256-904b-4d8d-8c98-d829582c6739