worker: Make it easier to use from 3rd party code
We split the child_process struct into two parts - one public half that
anyone implementing a worker will probably care about, and one private
one that adds statistics for nagios to use.

We add a simple callback arg to enter_worker, and make it public, so we
can override the callback in 3rd party code and make it do other stuff
with jobs than simply to spawn commands.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2305 5f96b256-904b-4d8d-8c98-d829582c6739