base/workers: Make it possible to register workers through query_handler
If a process opens the query handler socket, and sends a string on the
form
	@register_worker name=<name> plugin=<plugin1> plugin=<pluginN>
this process will be considered the only worker authorized for all the
plugins listed in the string.

It is possible for two processes to register for the same plugin, then
the load from the checks will be distributed between them evenly.

If the worker process dies, nagios will start to distribute the load
between the remaining workers for that plugin. If all of them dies,
nagios will move them back to the internal workers. Nagios will not
reschedule the already sent-out jobs, yet, though.

This creates the possibility to create workers that work completely
fork-free, which is sometimes a big win: embedded perl (or embedded
python/ruby/whatever) could be brought back through this API, rather
than through the old ifdef-heavy, in-core way. SNMP checks are another
example of checks that could benefit greatly from a dedicated worker.

Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2309 5f96b256-904b-4d8d-8c98-d829582c6739