lib/worker: Change delimiter so it doesn't clash with kvvec:PAIRSEP
Since kvvec:PAIRSEP is a nul byte (as all values must be C strings),
we run an unnecessary amount of memchr() calls for each worker
result we get back (and send).

We switch it to use "\1\0\0\0" instead (a nice 32-bit thing), which
we now have to copy explicitly to the end of the kvv-buffer.

Considering the fact that ascii code 1 is a non-printable character
and therefore should be highly unusual in any normal output sequence,
it should be pretty safe to use this.

An estimate is that we gain at least 10-15% performance from this
small change alone. Quite neat.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2316 5f96b256-904b-4d8d-8c98-d829582c6739