lib/nsock: Add nsock_printf_nul() as a separate function
The earlier nsock_printf() function would always nul-terminate its
strings, making it pretty hard to create strings that aren't meant
to end with a nul byte.

With this patch, that becomes positively trivial, and adding the
nul byte gets about four typed bytes trickier. None of the previous
callers need to be updated to use the new function.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2318 5f96b256-904b-4d8d-8c98-d829582c6739