lib/runcmd: Use _exit() instead of exit() when execvp() fails
Otherwise we'll run all atexit() handlers, which means profiling
worker processes when plugins are missing will cause the forked()
copy of the worker process try to merge profiling info with that
of earlier (failed) command executions. Since they lock the file
when merging and merging is non-trivial, that means we quickly
run out of available filedescriptors and end up in a severely
bad place, where checks are no longer run.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2329 5f96b256-904b-4d8d-8c98-d829582c6739