Add lib/nspath, the nagios path handling library
It's still pretty small but will suffice for our needs. The idea is
to easily and efficiently be able to resolve relative paths into
absolute ones, to better support moving ones installation around
in the filesystem (and, as an extension, make it easier for me to
run tests against other peoples configurations).

I haven't spent hours of effort to make this the most efficient
code in the world, since it will only be used when parsing the main
configuration, but it doesn't have to be and it seems plenty fast
enough for our purposes.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2335 5f96b256-904b-4d8d-8c98-d829582c6739