core: Allow relative paths for everything in nagios.cfg
Previously it was supported for object cfg_file and cfg_dir
directives, but that was inconsistent and surprising, since there
are lots of paths in nagios.cfg that can benefit from being relative
to the directory the config file resides in.

With this patch all paths in nagios.cfg are relative to the directory
where nagios.cfg resides (assuming I didn't miss any, ofcourse).

This makes it positively trivial to setup a test config, create
the directory structure for it and then send the entire thing to
a remote system for deployment or testing, without having to have
the exact same directory structure for everything.

While at it, we fix some broken indentation (in code we touched
upon anyway) and make check_result_path and temp_path two distinctly
different options, as opposed to earlier when check_result_path
would overwrite temp_path if check_result_path was specified last.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2336 5f96b256-904b-4d8d-8c98-d829582c6739