core: Make subsystems log with a unique tag
Since we now have subsystems (nerd and query-handler), it's kind of
important that we're able to find their logmessages quickly when
debugging, or that we can ignore them easily when machine-parsing
the logfiles. This patch makes queryhandler and nerd log messages
with 'qh: ' and 'nerd: ' prefixes, respectively, making it relatively
trivial to filter in or out their messages from the log.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2337 5f96b256-904b-4d8d-8c98-d829582c6739