core: Add timing points all the way up to the execution loop
This is necessary to (easily) measure where we spend a lot of time
while restarting, since config parsing is only a part of it. By
tracking it all the way until we enter the event execution loop
we effectively track all the time we use for anything at all
up until the point where we may be able to start running checks.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2338 5f96b256-904b-4d8d-8c98-d829582c6739