core: Launch workers by fork()+execv()
This solves the previously unsolvable problem of eventbroker modules
that allocate memory before our workers come alive, making our
workers huge compared to what we want them to be.

Before this patch, using 2954 hosts and 24850 services with Merlin
and livestatus loaded, workers use about 32mb resident memory, with
1.1mb shared.
With this patch, workers are down to about 2.1mb, with 720kb shared.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2348 5f96b256-904b-4d8d-8c98-d829582c6739