core: Add basic wproc accounting
It's good to know on a global scale how many workers are in which
state. This means we can now let all our desired workers connect
to us before we proceed to populate the scheduling queue and whatnot.

Basic timing point output shows it takes 0.0067 seconds to get six
workers online before we move on, which is clearly acceptable.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2351 5f96b256-904b-4d8d-8c98-d829582c6739