core: Test nagios_binary_path for executability after dropping privs
When we for some reason aren't allowed to execute ourselves as the
user we're supposed to run as, we'll be unable to spawn workers and
thus will be unable to do anything at all. With this patch we at
least tell the user (who obviously should read the logs) what we
tried (and failed) to do.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2353 5f96b256-904b-4d8d-8c98-d829582c6739