base/nerd: Stop using the 'nerd_sock'
The nerd_sock was only used during the (very) brief period when the
NERD radio existed and the query handler api did not, so it was
always just 0. That's a particularly bad value for a socket, since
it's a valid value as far as the iobroker is concerned, and that
could sometimes cause segfaults when deinitializing the nerd radio.

With this patch, we instead close all subscription sockets and
remove the nerd socket completely, which fixes the segfault.

Reported-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2355 5f96b256-904b-4d8d-8c98-d829582c6739