core: Maximize our RLIMIT_NOFILE and RLIMIT_NPROC values in workers
These are normally inherited by the root user otherwise, and now
that we have the capacity to spawn such an awful lot of checks in
a short period of time, we really do need to up these somewhat.

If we run out of processes, checks will start failing with EAGAIN
(Resource temporarily unavailable).
If we run out of filedescriptors, checks will start failing with
EMFILE (Too many open files). Neither of those are all that helpful,
so we'd best make sure to crank up the settings a bit.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2359 5f96b256-904b-4d8d-8c98-d829582c6739