core: Plug a minor memory leak and fix some invalid memory accesses
Accessing memory that's already free()'d leads to undefined behaviour.
If the free()'d memory is small enough and it's not claimed by some
other part of the program, that behaviour is usually the same as
acting as if the memory wasn't free()'d at all (ie, no sweat), but
it's retarded to rely on that, so we make sure to fix it.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2367 5f96b256-904b-4d8d-8c98-d829582c6739