xdata: Stop lying outrageously about precache savings
It's (sometimes) a little faster to use a precached config, but
reading the (often much) larger config takes a bit of time too,
and the savings are nowhere near as substantial as they once
were.

The largest saving I've seen is around 22%, but that's only for
certain parts of the restart sequence and not all of it, so it's
more like shaving 3-5% off of the total launch time.

However, since it's explicit which parts of config reading are
marked, we round it up to 25% when we give the hint.

Signed-off-by: Andreas Ericsson <ae@op5.se>

git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/trunk@2373 5f96b256-904b-4d8d-8c98-d829582c6739